+++ wxt.config.ts
@@ -23,6 +23,7 @@ export default defineConfig({
       "*://darts-downloads.peschi.org/*",
       "*://autodarts.x10.mx/*",
       "*://adt-socket.tobias-thiele.de/*",
+      "*://discord.com/api/webhooks/*",
     ],
     permissions: [
       "storage",