+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
-## [2.0.10] - 2025-04-2
+## [2.0.10] - 2025-04-1
 
 ### Added
 - Added keyboard shortcuts ('Y' to accept, 'N' to decline) for responding to lobby invitations
 
 ### Fixed
+- Fixed Discord webhooks permissions issue that was preventing proper webhook execution
 - Fixed a bug where match features did not apply after the first match when following a board
 
 ## [2.0.9] - 2025-04-1